pyproject.toml
[tool.poetry]
name = "dataviewer-mvp"
version = "0.1.0"
description = ""
authors = ["Ben Siderowf <bsiderowf@wustl.edu>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.1"
cantools = "^37.0.7"
python-dotenv = "^0.20.0"
Flask-SQLAlchemy = "^2.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
